Fix issue with obs crashing when exiting.
Fix obs crashing when exiting. Clean up more resources. Still have one memory leak but is minor since it only happens when loading the script. In general, this should not occur unless user is making lots of updates. Then it's just a one time hit. Will need to explore more to resolve this.